添加调用his门诊支付接口 添加调用his出院结算接口 添加调用his押金补缴接口

+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/his/controller/LeaveHosPayController.java
@@ -93,7 +93,7 @@ public class LeaveHosPayController extends BaseController
      * 修改出院结算
      */
     @GetMapping("/edit/{id}")
-    public String edit(@PathVariable("id") Integer id, ModelMap mmap)
+    public String edit(@PathVariable("id") Long id, ModelMap mmap)
     {
         LeaveHosPay leaveHosPay = leaveHosPayService.selectLeaveHosPayById(id);
         mmap.put("leaveHosPay", leaveHosPay);
